requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiomysql==0.3.2
aioodbc==0.5.0
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.12.1
asyncmy==0.2.11
asyncpg==0.30.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.0
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
numpy==2.4.3
openpyxl==3.1.5
oracledb==3.4.2
pandas==3.0.1
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pymssql==2.3.13
PyMySQL==1.1.2
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
requests==2.33.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.35
starlette==0.38.6
types-Jinja2==2.11.9
types-MarkupSafe==1.1.10
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarl==1.23.0
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0